Unify retreving info about gdlint instalation
Turns the method that look for gdlint's info and path into a single thing, still using private properties to cache the last found values and then making a single method that allows for re-checking for the path. Also puts a tool button for manually re-checking and add a optional file dialog popup for systems where the location of gdlint is not known.