Unify python command getting and further improvements path and version getting
Add a project setting for the path to python, as well as using that path everywhere it's should be.

Also adds a private method for parsing the version number that expected to be returned from gdlint, while also doing some very basic validation on that output (returns empty when invalid). Then it uses that method in the gdlint path finding method to improve the confidence that the automatically guessed gdlint command is really gdlint.

Also adds another possible way to call gdlint by using the module directly through python (`py -m gdtoolkit.linter`); though this method is checked it might never really apply as I did not change the the gdlint path finding function to also include potentially needed arguments, witch will likely cause this check to fail always and thus never use this new method. However, I see this method as the most consistent one as it interfaces directly with python itself so it can skip other more os specific options.